migration files updated

+++ modules/atlantis/forms/src/Module/Forms/Migrations/02_18_2016_130443_create_forms_config_table.php
@@ -14,20 +14,25 @@ class CreateFormsConfigTable extends Migration {
    * Run the migrations.
    */
   public function up() {
-
-    Schema::create('forms_config', function(\Illuminate\Database\Schema\Blueprint $table) {
-      $table->increments('id');
-      $table->string("config_key", 255);
-      $table->text("config_value")->nullable();
-      $table->timestamps();    
-    });
+    
+    if (!Schema::hasTable('forms_config')) {
+      Schema::create('forms_config', function(\Illuminate\Database\Schema\Blueprint $table) {
+        $table->increments('id');
+        $table->string("config_key", 255);
+        $table->text("config_value")->nullable();
+        $table->timestamps();
+      });
+    }
   }
 
   /**
    * Reverse the migrations.
    */
   public function down() {
-    Schema::drop('forms_config');
+    
+    if (Schema::hasTable('forms_config')) {
+      Schema::drop('forms_config');
+    }
   }
 
 }
