+++ modules/atlantis/forms/src/Module/Forms/Migrations/02_18_2016_130443_create_forms_results_table.php
@@ -15,24 +15,29 @@ class CreateFormsResultsTable extends Migration {
    */
   public function up() {
 
-    Schema::create('forms_results', function(\Illuminate\Database\Schema\Blueprint $table) {
-      $table->increments('id');
-      $table->integer("set_id");
-      $table->integer("form_id");
-      $table->text("field_label")->nullable();
-      $table->text("field_name")->nullable();
-      $table->text("field_value")->nullable();
-      $table->string("post_url", 255)->nullable();
-      $table->string("IP", 255)->nullable();
-      $table->timestamps();
-    });
+    if (!Schema::hasTable('forms_results')) {
+      Schema::create('forms_results', function(\Illuminate\Database\Schema\Blueprint $table) {
+        $table->increments('id');
+        $table->integer("set_id");
+        $table->integer("form_id");
+        $table->text("field_label")->nullable();
+        $table->text("field_name")->nullable();
+        $table->text("field_value")->nullable();
+        $table->string("post_url", 255)->nullable();
+        $table->string("IP", 255)->nullable();
+        $table->timestamps();
+      });
+    }
   }
 
   /**
    * Reverse the migrations.
    */
   public function down() {
-    Schema::drop('forms_results');
+
+    if (Schema::hasTable('forms_results')) {
+      Schema::drop('forms_results');
+    }
   }
 
 }
