+++ modules/atlantis/forms/src/Module/Forms/Migrations/2016_06_15_172722_add_redirect_error.php
@@ -4,27 +4,28 @@ use Illuminate\Database\Migrations\Migration;
 
 class AddRedirectError extends Migration {
 
-        /**
-        * Run the migrations.
-        */
-        public function up()
-        {
-                //
-                Schema::table('forms', function(\Illuminate\Database\Schema\Blueprint $table)
-                {                   
-                        $table->string("redirect_url_error", 255)->nullable();
-                });
-        }
+  /**
+   * Run the migrations.
+   */
+  public function up() {
 
-        /**
-        * Reverse the migrations.
-        */
-        public function down()
-        {
-                 Schema::table('forms', function(\Illuminate\Database\Schema\Blueprint $table)
-                {                   
-                        $table->dropColumn("redirect_url_error");
-                });
-        }
+    if (!Schema::hasColumn('forms', 'redirect_url_error')) {
+      Schema::table('forms', function(\Illuminate\Database\Schema\Blueprint $table) {
+        $table->string("redirect_url_error", 255)->nullable();
+      });
+    }
+  }
+
+  /**
+   * Reverse the migrations.
+   */
+  public function down() {
+    
+    if (Schema::hasColumn('forms', 'redirect_url_error')) {
+      Schema::table('forms', function(\Illuminate\Database\Schema\Blueprint $table) {
+        $table->dropColumn("redirect_url_error");
+      });
+    }
+  }
 
 }
