+++ modules/atlantis/menus/src/Module/Menus/Migrations/2015_12_15_000006_create_menu_caches_table.php
@@ -8,20 +8,25 @@ class CreateMenuCachesTable extends Migration {
    * Run the migrations.
    */
   public function up() {
-    //
-    Schema::create('menu_caches', function(\Illuminate\Database\Schema\Blueprint $table) {
-      $table->increments('id');
-      $table->integer('menu_id')->nullable();
-      $table->text('compiled');
-      $table->timestamps();
-    });
+
+    if (!Schema::hasTable('menu_caches')) {
+      Schema::create('menu_caches', function(\Illuminate\Database\Schema\Blueprint $table) {
+        $table->increments('id');
+        $table->integer('menu_id')->nullable();
+        $table->text('compiled');
+        $table->timestamps();
+      });
+    }
   }
 
   /**
    * Reverse the migrations.
    */
   public function down() {
-    Schema::drop('menu_caches');
+
+    if (!Schema::hasTable('menu_caches')) {
+      Schema::drop('menu_caches');
+    }
   }
 
 }
