+++ modules/atlantis/menus/src/Module/Menus/Migrations/2015_12_15_000006_create_menus_table.php
@@ -8,22 +8,27 @@ class CreateMenusTable extends Migration {
    * Run the migrations.
    */
   public function up() {
-    //
-    Schema::create('menus', function(\Illuminate\Database\Schema\Blueprint $table) {
-      $table->increments('id');
-      $table->string("name", 255)->nullable();
-      $table->text('menu_attributes');
-      $table->string("css", 255)->nullable();
-      $table->string("element_id", 255)->nullable();
-      $table->timestamps();
-    });
+
+    if (!Schema::hasTable('menus')) {
+      Schema::create('menus', function(\Illuminate\Database\Schema\Blueprint $table) {
+        $table->increments('id');
+        $table->string("name", 255)->nullable();
+        $table->text('menu_attributes');
+        $table->string("css", 255)->nullable();
+        $table->string("element_id", 255)->nullable();
+        $table->timestamps();
+      });
+    }
   }
 
   /**
    * Reverse the migrations.
    */
   public function down() {
-    Schema::drop('menus');
+
+    if (Schema::hasTable('menus')) {
+      Schema::drop('menus');
+    }
   }
 
 }
