+++ modules/atlantis/openweather/src/Module/OpenWeather/Migrations/03_07_2016_184450_create_openweather_table.php
@@ -15,19 +15,24 @@ class CreateOpenweatherTable extends Migration {
    */
   public function up() {
 
-    Schema::create('openweather', function(\Illuminate\Database\Schema\Blueprint $table) {
-      $table->increments('id');
-      $table->string("app_id");
-      $table->string("temperature")->default('C');
-      $table->timestamps();
-    });
+    if (!Schema::hasTable('openweather')) {
+      Schema::create('openweather', function(\Illuminate\Database\Schema\Blueprint $table) {
+        $table->increments('id');
+        $table->string("app_id");
+        $table->string("temperature")->default('C');
+        $table->timestamps();
+      });
+    }
   }
 
   /**
    * Reverse the migrations.
    */
   public function down() {
-    Schema::drop('openweather');
+
+    if (Schema::hasTable('openweather')) {
+      Schema::drop('openweather');
+    }
   }
 
 }
